+++ bin/imc_desafio_dio.dart
@@ -1,5 +1,5 @@
-import 'package:imc_desafio_dio/imc_desafio_dio.dart' as imc_desafio_dio;
+import 'package:imc_desafio_dio/app.dart' as app;
 
 void main(List<String> arguments) {
-  print('Hello world: ${imc_desafio_dio.calculate()}!');
+  app.execute();
 }