+++ test/imc_desafio_dio_test.dart
@@ -1,8 +1,11 @@
-import 'package:imc_desafio_dio/imc_desafio_dio.dart';
+import 'package:imc_desafio_dio/imc.dart';
 import 'package:test/test.dart';
 
 void main() {
   test('calculate', () {
-    expect(calculate(), 42);
+    expect(IMC.calculate(peso: 57.8, altura: 158), 23.15);
+  });
+  test('classificacao do imc', () {
+    expect(IMC.imcClassificacao(23.15), 'Saudavel');
   });
 }