+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 1.0.0
+## 1.0.0 2023-08-21
 
 - Initial version.
+
+## 1.0.1 2023-08-21
+
+- Calculo do imc
+- Classificacao do imc
+
+## 1.0.2 2023-08-21
+
+- Interacao com o usuario pelo terminal
+- tratamento de erros